pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "httpx>=0.28.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
